+++ src/datastructures/lts_queue.h
@@ -164,7 +164,7 @@ class LTSQueue : Queue<T>{
           scal::ThreadLocalAllocator::Get().CallocAligned(num_threads_, sizeof(std::atomic<SPBuffer*>), 
             scal::kCachePrefetch * 4));
 
-      for (int i = 0; i < num_threads_; i++) {
+      for (uint64_t i = 0; i < num_threads_; i++) {
          spBuffers_[i].store(NULL); 
       }
       
@@ -223,7 +223,7 @@ class LTSQueue : Queue<T>{
       uint64_t sum1 = 0;
       uint64_t sum2 = 1;
 
-      for (int i = 0; i < num_threads_; i++) {
+      for (uint64_t i = 0; i < num_threads_; i++) {
         sum1 += *counter1_[i];
         sum2 += *counter2_[i];
       }
