+++ src/datastructures/ts_queue_buffer.h
@@ -148,7 +148,7 @@ class TSQueueBuffer {
           scal::ThreadLocalAllocator::Get().CallocAligned(num_threads_, sizeof(std::atomic<SPBuffer*>), 
             scal::kCachePrefetch * 4));
 
-      for (int i = 0; i < num_threads_; i++) {
+      for (uint64_t i = 0; i < num_threads_; i++) {
          spBuffers_[i].store(NULL); 
       }
       
@@ -176,7 +176,7 @@ class TSQueueBuffer {
           scal::ThreadLocalAllocator::Get().CallocAligned(num_threads_, sizeof(Item**), 
             scal::kCachePrefetch * 4));
 
-      for (int i = 0; i < num_threads_; i++) {
+      for (uint64_t i = 0; i < num_threads_; i++) {
 
         emptiness_check_pointers_[i] = static_cast<Item**> (
             scal::ThreadLocalAllocator::Get().CallocAligned(num_threads_, sizeof(Item*), 
@@ -212,7 +212,7 @@ class TSQueueBuffer {
       uint64_t sum1 = 0;
       uint64_t sum2 = 1;
 
-      for (int i = 0; i < num_threads_; i++) {
+      for (uint64_t i = 0; i < num_threads_; i++) {
         sum1 += *counter1_[i];
         sum2 += *counter2_[i];
       }
