+++ src/datastructures/ts_stack_buffer.h
@@ -196,7 +196,7 @@ class TSStackBuffer {
             scal::ThreadLocalAllocator::Get().CallocAligned(num_threads_, sizeof(SPBuffer), 
               scal::kCachePrefetch * 4));
 
-       for (int i = 0; i < num_threads_; i++) {
+       for (uint64_t i = 0; i < num_threads_; i++) {
          spBuffers_[i].store(NULL); 
 
         emptiness_check_pointers_[i] = static_cast<Item**> (
@@ -254,7 +254,7 @@ class TSStackBuffer {
       uint64_t sum1 = 0;
       uint64_t sum2 = 1;
 
-      for (int i = 0; i < num_threads_; i++) {
+      for (uint64_t i = 0; i < num_threads_; i++) {
         sum1 += *counter1_[i];
         sum2 += *counter2_[i];
       }
